refactor to avoid import cycle once analysis change lands
[ci skip-rust]
