+++ src/python/pants/backend/experimental/go/register.py
@@ -16,6 +16,7 @@ from pants.backend.go.target_types import (
 from pants.backend.go.util_rules import (
     assembly,
     build_pkg,
+    build_pkg_target,
     first_party_pkg,
     go_mod,
     import_analysis,
@@ -34,6 +35,7 @@ def rules():
     return [
         *assembly.rules(),
         *build_pkg.rules(),
+        *build_pkg_target.rules(),
         *check.rules(),
         *third_party_pkg.rules(),
         *golang.rules(),
