+++ src/python/pants/backend/go/goals/package_binary.py
@@ -9,7 +9,8 @@ from pants.backend.go.target_types import (
     GoBinaryMainPackageField,
     GoBinaryMainPackageRequest,
 )
-from pants.backend.go.util_rules.build_pkg import BuildGoPackageTargetRequest, BuiltGoPackage
+from pants.backend.go.util_rules.build_pkg import BuiltGoPackage
+from pants.backend.go.util_rules.build_pkg_target import BuildGoPackageTargetRequest
 from pants.backend.go.util_rules.import_analysis import ImportConfig, ImportConfigRequest
 from pants.backend.go.util_rules.link import LinkedGoBinary, LinkGoBinaryRequest
 from pants.core.goals.package import (
