+++ src/python/pants/backend/go/util_rules/assembly_integration_test.py
@@ -16,6 +16,7 @@ from pants.backend.go.target_types import GoBinaryTarget, GoModTarget
 from pants.backend.go.util_rules import (
     assembly,
     build_pkg,
+    build_pkg_target,
     first_party_pkg,
     go_mod,
     import_analysis,
@@ -39,6 +40,7 @@ def rule_runner() -> RuleRunner:
             *import_analysis.rules(),
             *package_binary.rules(),
             *build_pkg.rules(),
+            *build_pkg_target.rules(),
             *first_party_pkg.rules(),
             *go_mod.rules(),
             *link.rules(),
